display user if connected
