+++ template/src/main/twirl/ch.epfl.scala.index.views/editproject.scala.html
@@ -1,6 +1,6 @@
-@import ch.epfl.scala.index.model.{Project, Descending, Artifact, SemanticVersion}
-@(project: Project, artifact: Artifact.Reference, version: Option[SemanticVersion], production: Boolean)
-@main(title = "Home", production) {
+@import ch.epfl.scala.index.model._
+@(project: Project, artifact: Artifact.Reference, version: Option[SemanticVersion], user: Option[UserInfo], production: Boolean)
+@main(title = s"Edit ${artifact.name}", showSearch = true, searchQuery = "", user, production) {
   <main id="container-project">
 
     @headproject(project, artifact, version)
