+++ template/src/main/twirl/ch.epfl.scala.index.views/headproject.scala.html
@@ -1,4 +1,4 @@
-@import ch.epfl.scala.index.model.{Project, Descending, Artifact, SemanticVersion}
+@import ch.epfl.scala.index.model._
 @(project: Project, artifactRef: Artifact.Reference, version: Option[SemanticVersion])
 <div class="head-project">
   <div class="container">
