+++ template/src/main/twirl/ch.epfl.scala.index.views/paginate.scala.html
@@ -1,4 +1,4 @@
-@import ch.epfl.scala.index.model.Pagination
+@import ch.epfl.scala.index.model._
 @(pagination: Pagination, uri: String)
 
 <nav>
