+++ template/src/main/twirl/ch.epfl.scala.index.views/searchresult.scala.html
@@ -1,8 +1,7 @@
-@import ch.epfl.scala.index._
-@import ch.epfl.scala.index.model.{Project, Pagination, Url}
+@import ch.epfl.scala.index.model._
 
-@(query: String, sorting: Option[String], pagination: Pagination, projects: List[Project], production: Boolean)
-@main(title = "Home", production, searchQuery = query) {
+@(query: String, sorting: Option[String], pagination: Pagination, projects: List[Project], user: Option[UserInfo], production: Boolean)
+@main(title = s"Search $query", showSearch = false, searchQuery = "", user, production) {
 <main id="container-search">
   <div class="container">
     @if(pagination.total != 0 ) {
