+++ webapp/shared/src/main/scala/ch.epfl.scala.index.api/Api.scala
@@ -1,15 +1,11 @@
 package ch.epfl.scala.index
+package api
 
-// import model._
+import model._
+
 import scala.concurrent.Future
 
-package api {
-  case class UserInfo(login: String, name: String, avatarUrl: String) {
-    def sizedAvatarUrl(size: Int) = avatarUrl + "&s" + size.toString
-  }
-
-  trait Api {
-    def userInfo(): Option[UserInfo]
-    def autocomplete(q: String): Future[List[(String, String, String)]]
-  }
+trait Api {
+  def userInfo(): Option[UserInfo]
+  def autocomplete(q: String): Future[List[(String, String, String)]]
 }
